Added basic enemy shooting
it "works" the only problem is that with my enemy owning guns system i havent been able to figure out a way for the gun to recognize its being held by a player or not.
Oh, by the way, the way this works is that the enemy gets a gun, and that gun shoots bullets. All three of those are seperate nodes so they can mix and match